requirements.txt
python-dotenv
rt
requests
pyyaml